LPS-54677 1. Move to a proper folder 2. Add configuration and build files 3. Add interfaces to the module 4. Remove registry and static classes
