+++ modules/portal/portal-expression/src/com/liferay/portal/expression/internal/Variable.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.expression;
+package com.liferay.portal.expression.internal;
 
 /**
  * @author Miguel Angelo Caldas Gallindo
