+++ modules/portal/portal-expression/test/unit/com/liferay/portal/expression/VariableNamesExtractorTest.java
@@ -14,7 +14,7 @@
 
 package com.liferay.portal.expression;
 
-import com.liferay.portal.kernel.expression.VariableNamesExtractor;
+import com.liferay.portal.expression.internal.VariableNamesExtractor;
 
 import java.util.List;
